Added Audit History API
