+++ search-api/search-service/app/managers/SearchOperations.java
@@ -1,5 +1,5 @@
 package managers;
 
 public enum SearchOperations {
-    INDEX_SEARCH, LANGUAGE_SEARCH, METRICS, COUNT, GROUP_SEARCH_RESULT_BY_OBJECTTYPE, MULTI_LANGUAGE_WORD_SEARCH, MULTI_LANGUAGE_SYNSET_SEARCH, HEALTH
+    INDEX_SEARCH, SEARCH_OPERATION_AND,LANGUAGE_SEARCH, METRICS, COUNT, GROUP_SEARCH_RESULT_BY_OBJECTTYPE, MULTI_LANGUAGE_WORD_SEARCH, MULTI_LANGUAGE_SYNSET_SEARCH, HEALTH
 }
