+++ search-api/search-service/app/modules/SearchModule.scala
@@ -1,7 +1,7 @@
 package modules
 
 import com.google.inject.AbstractModule
-import org.sunbird.actors.{HealthActor, SearchActor}
+import org.sunbird.actors.{AuditHistoryActor, HealthActor, SearchActor}
 import org.sunbird.telemetry.TelemetryGenerator
 import play.libs.akka.AkkaGuiceSupport
 import utils.ActorNames
@@ -12,6 +12,7 @@ class SearchModule extends AbstractModule with AkkaGuiceSupport {
         super.configure()
         bindActor(classOf[HealthActor], ActorNames.HEALTH_ACTOR)
         bindActor(classOf[SearchActor], ActorNames.SEARCH_ACTOR)
+        bindActor(classOf[AuditHistoryActor], ActorNames.AUDIT_HISTORY_ACTOR)
         TelemetryGenerator.setComponent("search-service")
         println("Initialized application actors for search-service")
     }
