+++ search-api/search-service/app/utils/ActorNames.scala
@@ -4,5 +4,6 @@ object ActorNames {
 
     final val HEALTH_ACTOR = "healthActor"
     final val SEARCH_ACTOR = "SearchActor"
+    final val AUDIT_HISTORY_ACTOR = "AuditHistoryActor"
 
 }
