+++ search-api/search-service/conf/application.conf
@@ -313,3 +313,5 @@ ekstepPlatformApiUserId="search-service"
 
 content.tagging.property=["subject","medium"]
 search.payload.log_enable=true
+
+compositesearch.index.name="kp_audit_log"
